atualizacoes de path das páginas create, delete, put e delete(barradas pelo CORS até o momento)
